Add files via upload
This Perl script is used to compare two chromosomes (A and B) or chromosomal segments using blast. For this chromosome A is split into segments of a user specified size (with sliding window), which are then blasted against a blast database of chromosome B. Note, that only the best blast hit of chromosome A segment_X on chromosome B is reported.

Thank you for using WICKERsoft!